Make channel flags part of the Channel class
